+++ Example/LYToolsKit/LYViewController.m
@@ -7,6 +7,7 @@
 //
 
 #import "LYViewController.h"
+#import <LYToolsKit/LYToolsKit.h>
 
 @interface LYViewController ()
 
@@ -17,7 +18,19 @@
 - (void)viewDidLoad
 {
     [super viewDidLoad];
-	// Do any additional setup after loading the view, typically from a nib.
+    UIBarButtonItem *btn = [UIBarButtonItem ly_itemWithNormalImageName:@"消息中心-系统" highImageName:@"消息中心-系统" target:self action:@selector(btnClick)];
+    self.navigationItem.rightBarButtonItem = btn;
+
+    // Do any additional setup after loading the view, typically from a nib.
+    LYHyperlinksButton *btn1 = [LYHyperlinksButton ly_ViewWithColor:[UIColor redColor]];
+    [btn1 setTitle:@"啧啧啧" forState:UIControlStateNormal];
+    btn1.frame = CGRectMake(50, 100, 50, 50);
+    
+    [self.view addSubview:btn1];
+}
+
+- (void)btnClick {
+    
 }
 
 - (void)didReceiveMemoryWarning