+++ LYToolsKit/Utility/UIButton/LYHyperlinksButton.h
@@ -1,0 +1,13 @@
+//
+//  LYHyperlinksButton.h
+//  shuaidanbao
+//
+//  Created by 吴浪 on 15/10/20.
+//  Copyright © 2015年 sdb. All rights reserved.
+//  超链接按钮
+
+#import <UIKit/UIKit.h>
+
+@interface LYHyperlinksButton : UIButton
+@property (strong, nonatomic) UIColor *lineColor; /**< 下划线颜色 */
+@end